Better support for MOF-74-like structures
Adding adsorbates to five-coordinate, pseudo square pyramidal geometries should be in the same position regardless of the position of the nonplanar 5th adsorbate (e.g. compare MOF-74 with M2(DSBDC)). Now, MAI will cycle through every cnum-1 combo of coordinating atoms and see if there is a planar set. If there is, only take the adsorbate with respect to the planar set of coordinates, not the sum of distance vectors.